rename test.py -> test_all.py for pytest
